S10P22A407-10 feat: Feat Custom Read

+++ backend/api/src/main/java/store/dalkak/api/custom/service/CustomServiceImpl.java
@@ -3,32 +3,48 @@ package store.dalkak.api.custom.service;
 import jakarta.transaction.Transactional;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 import store.dalkak.api.cocktail.domain.Cocktail;
 import store.dalkak.api.cocktail.domain.ingredient.Ingredient;
 import store.dalkak.api.cocktail.domain.ingredient.Unit;
+import store.dalkak.api.cocktail.dto.CocktailDto;
+import store.dalkak.api.cocktail.exception.CocktailErrorCode;
+import store.dalkak.api.cocktail.exception.CocktailException;
 import store.dalkak.api.cocktail.repository.CocktailRepository;
 import store.dalkak.api.cocktail.repository.UnitRepository;
 import store.dalkak.api.cocktail.repository.ingredient.IngredientRepository;
 import store.dalkak.api.custom.domain.Custom;
 import store.dalkak.api.custom.domain.CustomIngredient;
+import store.dalkak.api.custom.dto.CustomCocktailDto;
+import store.dalkak.api.custom.dto.CustomIngredientDetailDto;
 import store.dalkak.api.custom.dto.CustomIngredientDto;
 import store.dalkak.api.custom.dto.CustomIngredientModifyDto;
 import store.dalkak.api.custom.dto.CustomModifyDto;
 import store.dalkak.api.custom.dto.request.CustomCreateReqDto;
+import store.dalkak.api.custom.dto.response.CustomDetailResDto;
+import store.dalkak.api.custom.dto.response.CustomListResDto;
+import store.dalkak.api.custom.exception.CustomErrorCode;
+import store.dalkak.api.custom.exception.CustomException;
 import store.dalkak.api.custom.repository.CustomIngredientRepository;
 import store.dalkak.api.custom.repository.CustomRepository;
 import store.dalkak.api.global.config.ImageConfig;
 import store.dalkak.api.user.domain.Member;
-import store.dalkak.api.user.repository.MemberRepository;
 import store.dalkak.api.user.dto.MemberDto;
+import store.dalkak.api.user.dto.UserDto;
+import store.dalkak.api.user.exception.UserErrorCode;
+import store.dalkak.api.user.exception.UserException;
+import store.dalkak.api.user.repository.MemberRepository;
 
 @Service
 @RequiredArgsConstructor
 @Slf4j
+@Transactional
 public class CustomServiceImpl implements CustomService {
 
     private final ImageConfig imageConfig;
@@ -46,7 +62,6 @@ public class CustomServiceImpl implements CustomService {
     private final MemberRepository memberRepository;
 
     @Override
-    @Transactional
     public void createCustomCocktail(MultipartFile image, CustomCreateReqDto customCreateReqDto,
         MemberDto memberDto) {
         String imageUrl = imageConfig.uploadImage(image);
@@ -70,9 +85,13 @@ public class CustomServiceImpl implements CustomService {
     }
 
     @Override
-    @Transactional
-    public void deleteCustomCocktail(Long customCocktailId) {
+    public void deleteCustomCocktail(Long userId, Long customCocktailId) {
         Custom custom = customRepository.findCustomById(customCocktailId);
+
+        if(!Objects.equals(userId, custom.getMember().getId())) {
+            throw new UserException(UserErrorCode.FORBIDDEN);
+        }
+
         // 이미지 삭제
         imageConfig.deleteImage(custom.getImage());
 
@@ -86,10 +105,15 @@ public class CustomServiceImpl implements CustomService {
     }
 
     @Override
-    @Transactional
-    public void modifyCustomCocktail(Long customCocktailId, MultipartFile image,
+    public void modifyCustomCocktail(Long userId, Long customCocktailId, MultipartFile image,
+
         CustomCreateReqDto customCreateReqDto) {
+
         Custom custom = customRepository.findCustomById(customCocktailId);
+
+        if(!Objects.equals(userId, custom.getMember().getId())) {
+            throw new UserException(UserErrorCode.FORBIDDEN);
+        }
 
         // 현재 커스텀 칵테일이 저장하고 있는 커스텀 재료 아이디를 리스트에 저장
         List<Long> customIngredientIdList = new ArrayList<>();
@@ -115,7 +139,7 @@ public class CustomServiceImpl implements CustomService {
 
         String imageUrl;
         // 만약 이미지를 변경했다면
-        if(image != null) {
+        if (image != null) {
             imageConfig.deleteImage(custom.getImage());
             imageUrl = imageConfig.uploadImage(image);
         } else {
@@ -127,5 +151,63 @@ public class CustomServiceImpl implements CustomService {
             customCreateReqDto.getCustomRecipe(), imageUrl, customCreateReqDto.getOpen());
 
         customRepository.modifyCustomCocktail(customCocktailId, customModifyDto);
+    }
+
+    @Override
+    public CustomListResDto getCustomList(Long cocktailId, Pageable page) {
+        Cocktail targetcocktail = cocktailRepository.findById(cocktailId)
+            .orElseThrow(() -> new CocktailException(
+                CocktailErrorCode.FAIL_TO_FIND_COCKTAIL));
+        Page<Custom> customPage = customRepository.findAllByCocktailOrderByIdDesc(targetcocktail,
+            page);
+
+        return CustomListResDto.builder()
+            .customCocktails(toCustomCocktailDtoList(customPage.getContent()))
+            .currentPage(customPage.getPageable().getPageNumber() + 1)
+            .totalPage(customPage.getTotalPages())
+            .totalElements((customPage.getTotalElements()))
+            .build();
+    }
+
+    @Override
+    public CustomDetailResDto findCustom(Long customCocktailId) {
+        Custom targetCustom = customRepository.findById(customCocktailId).orElseThrow(
+            () -> new CustomException(CustomErrorCode.FAIL_TO_FIND_CUSTOM));
+        UserDto user = new UserDto(targetCustom.getMember().getId(),
+            targetCustom.getMember().getNickname());
+        CocktailDto cocktail = new CocktailDto(targetCustom.getCocktail().getId(),
+            targetCustom.getCocktail().getName(), targetCustom.getCocktail().getKrName(),
+            targetCustom.getCocktail().getImage(), targetCustom.getCocktail().getHeartCount());
+        List<CustomIngredient> customIngredients = customIngredientRepository.findAllByCustom(
+            targetCustom);
+        List<CustomIngredientDetailDto> customIngredientDtoList = new ArrayList<>();
+        for (CustomIngredient customIngredient : customIngredients) {
+            customIngredientDtoList.add(new CustomIngredientDetailDto(
+                customIngredient.getIngredient().getId(),
+                customIngredient.getIngredient().getName(),
+                customIngredient.getIngredient().getImage(),
+                customIngredient.getAmount(),
+                customIngredient.getUnit()));
+        }
+
+        return CustomDetailResDto.of(targetCustom, user, cocktail, customIngredientDtoList);
+    }
+
+
+    private List<CustomCocktailDto> toCustomCocktailDtoList(List<Custom> customs) {
+        List<CustomCocktailDto> customCocktailDtoList = new ArrayList<>();
+        for (Custom custom : customs) {
+            customCocktailDtoList
+                .add(CustomCocktailDto
+                    .builder()
+                    .id(custom.getId())
+                    .image(custom.getCocktail().getImage())
+                    .name(custom.getName())
+                    .summary(custom.getSummary())
+                    .user(UserDto.builder().id(custom.getMember().getId())
+                        .nickname(custom.getMember().getNickname()).build())
+                    .build());
+        }
+        return customCocktailDtoList;
     }
 }
